% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend}
\alias{detrend}
\title{Fast detrend}
\usage{
detrend(x)
}
\arguments{
\item{x}{numeric vector to be detrended.}
}
\value{
A numeric vector of the same length as x with detrended values.
}
\description{
Removes a linear trend.
}
\examples{
x <- rnorm(10) + 1:10

detrend(x)

}
