% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HADSST}
\alias{HADSST}
\title{HadSST data}
\usage{
HADSST()
}
\description{
Returns the location of the HadSST file location, downloading the data from
the MetOffice website if needed.
}
\references{
Rayner, N. A.; Parker, D. E.; Horton, E. B.; Folland, C. K.; Alexander, L. V.; Rowell, D. P.; Kent, E. C.; Kaplan, A. (2003) Global analyses of sea surface temperature, sea ice, and night marine air temperature since the late nineteenth century J. Geophys. Res.Vol. 108, No. D14, 4407 10.1029/2002JD002670
}
