% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaluate.R
\name{pvaluate}
\alias{pvaluate}
\title{Compute p.values}
\usage{
pvaluate(estimate, std.error, df, adjustment = "none")
}
\arguments{
\item{estimate}{estimate of the regression (or other fit)}

\item{std.error}{standard error}

\item{df}{degrees of freedom}

\item{adjustment}{method for adjustment, see \code{\link[stats:p.adjust]{stats::p.adjust()}}.}
}
\description{
Computes p.values from the result of \code{\link[metR:FitLm]{metR::FitLm()}} (or any assumed student-t-distributed
statistic) with optional adjustments.
}
