% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{quick_map}
\alias{quick_map}
\title{Simple map for ggplot2}
\usage{
quick_map(
  fill = "white",
  color = "black",
  size = 0.2,
  ticks.lat = 10,
  ylim = c(-55, -20)
)
}
\arguments{
\item{fill, color, size}{parameters passed to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}.}

\item{ticks.lat}{sepaeration between latitude marks}

\item{ylim}{latitude limites passed to \code{\link[ggplot2:coord_map]{ggplot2::coord_quickmap()}}}
}
\description{
A \code{geom_polygon()} with x and y scales for longitude and latitude and quickmap
coordinates.
}
